filter, filter category, tag, tag group modules

+++ Implementation/Services/FilterService.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using IST.Interfaces.IServices;
 using IST.Interfaces.Repository;
 using IST.Models.DomainModels;
@@ -34,16 +36,21 @@ namespace IST.Implementation.Services
             return filterRepository.Find(filterId);
         }
         
-        public bool SaveFilter(Filter filter)
+        public bool SaveOrUpdateFilter(Filter filter)
         {
-            filterRepository.Add(filter);
-            filterRepository.SaveChanges();
-            return true;
-        }
-
-        public bool UpdateFilter(Filter filter)
-        {
-            filterRepository.Update(filter);
+            if (filter.Id > 0)
+            {
+                var filterToUpdate = filterRepository.Find(filter.Id);
+                filterToUpdate.DisplayValue = filter.DisplayValue;
+                filterToUpdate.FilterCategoryId = filter.FilterCategoryId;
+                filterToUpdate.RecLastUpdatedById = filter.RecLastUpdatedById;
+                filterToUpdate.RecLastUpdatedOn = filter.RecLastUpdatedOn;
+                filterRepository.Update(filterToUpdate);
+            }
+            else
+            {
+                filterRepository.Add(filter);
+            }
             filterRepository.SaveChanges();
             return true;
         }
@@ -51,6 +58,10 @@ namespace IST.Implementation.Services
         public bool DeleteFilter(int filterId)
         {
             var toDelete = filterRepository.Find(filterId);
+            if (toDelete.Solutions.Any())
+            {
+                throw new Exception("You cannot delete " + toDelete.DisplayValue + " as it is being used in Solutions.");
+            }
             filterRepository.Delete(toDelete);
             filterRepository.SaveChanges();
             return true;
