+++ Implementation/Services/TagGroupService.cs
@@ -1,4 +1,6 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using IST.Interfaces.IServices;
 using IST.Interfaces.Repository;
 using IST.Models.DomainModels;
@@ -34,11 +36,11 @@ namespace IST.Implementation.Services
         {
             if (tagGroup.Id > 0)
             {
-                var categoryToUpdate = tagGroupRepository.Find(tagGroup.Id);
-                categoryToUpdate.DisplayValue = tagGroup.DisplayValue;
-                categoryToUpdate.RecLastUpdatedById = tagGroup.RecLastUpdatedById;
-                categoryToUpdate.RecLastUpdatedOn = tagGroup.RecLastUpdatedOn;
-                tagGroupRepository.Update(categoryToUpdate);
+                var groupToUpdate = tagGroupRepository.Find(tagGroup.Id);
+                groupToUpdate.DisplayValue = tagGroup.DisplayValue;
+                groupToUpdate.RecLastUpdatedById = tagGroup.RecLastUpdatedById;
+                groupToUpdate.RecLastUpdatedOn = tagGroup.RecLastUpdatedOn;
+                tagGroupRepository.Update(groupToUpdate);
             }
             else
             {
@@ -51,6 +53,10 @@ namespace IST.Implementation.Services
         public bool DeleteTagGroup(int tagGroupId)
         {
             var tagGroupToDelete = tagGroupRepository.Find(tagGroupId);
+            if (tagGroupToDelete.Tags.Any())
+            {
+                throw new Exception("You cannot delete " + tagGroupToDelete.DisplayValue + " as it is being used in Tag.");
+            }
             tagGroupRepository.Delete(tagGroupToDelete);
             tagGroupRepository.SaveChanges();
             return true;
